requirements.txt
pywin32>=306
customtkinter>=5.2.0
pillow>=10.0.0
PyPDF2>=3.0.0

